refactor(utils): colocate constructor types with extend
